requirements.txt
azure-devops==7.1.0b3
msrest>=0.6.0
python-dotenv>=1.0.0

